backend/requirements.txt
autopep8==1.5.7
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
decorator==5.0.7
elasticsearch==7.12.1
fast-autocomplete==0.7.1
geographiclib==1.50
geopy==2.1.0
google==3.0.0
google-api-core==1.30.0
google-auth==1.32.0
google-cloud-core==1.7.1
google-cloud-translate==3.2.1
googleapis-common-protos==1.53.0
googletrans==4.0.0rc1
greenlet==1.1.0
grpcio==1.38.1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jdatetime==3.6.2
jedi==0.18.0
lxml==4.6.3
msgpack==1.0.2
numpy==1.22.3
packaging==20.9
pandas==1.2.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prettytable==2.1.0
prompt-toolkit==3.0.18
proto-plus==1.18.1
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
Pygments==2.8.1
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rfc3986==1.5.0
rsa==4.7.2
signalrcore==0.9.2
six==1.15.0
sniffio==1.2.0
soupsieve==2.2.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
websocket-client==0.54.0
